+++ tests/native/test_single_client.py
@@ -140,6 +140,12 @@ class SSH2ClientTest(SSH2TestCase):
         finally:
             os.rmdir('adir')
 
+    def test_pkey_from_memory(self):
+        with open(self.user_key, 'rb') as fh:
+            key_data = fh.read()
+        SSHClient(self.host, port=self.port,
+                  pkey=key_data, num_retries=1, timeout=1)
+
     def test_execute(self):
         host_out = self.client.run_command(self.cmd)
         output = list(host_out.stdout)