+++ Changelog.rst
@@ -1,6 +1,18 @@
 Change Log
 ============
 
+2.8.0
++++++
+
+Changes
+--------
+
+* All clients now support private key data as bytes in ``pkey`` parameter for authentication from in-memory private key
+  data - #
+* Parallel clients now read a provided private key path only once and use in-memory data for authentication to avoid
+  reading same file multiple times, if a path is provided.
+
+
 2.7.1
 +++++